Crate test_thread_local

Offers thread-local storage (TLS) in Theseus, or rather, task-local storage.

This crate offers fully-fledged support for fast access to and dynamic allocation of arbitrary objects that will exist on a per-task (per-“thread”) basis. It integrates with Rust’s #[thread_local] attribute to offer features similar to that of the Rust standard library, e.g., the thread_local!() macro, in a more ergonomic way more friendly to struct composition.

Structs

Complex

MyStruct

Statics

COMPLEX

LOCAL_U8_1

LOCAL_U8_2

LOCAL_U16

LOCAL_USIZE

LOCAL_ZERO

MY_STRUCT

Functions

main

test_tls